+++ sdk/sql/azure-resourcemanager-sql/src/main/java/com/azure/resourcemanager/sql/implementation/SqlDatabaseExportRequestImpl.java
@@ -158,6 +158,16 @@ public class SqlDatabaseExportRequestImpl extends ExecutableImpl<SqlDatabaseImpo
         return this.withLoginAndPassword(administratorLogin, administratorPassword);
     }
 
+    @Override
+    public SqlDatabaseExportRequestImpl withManagedIdentity(String managedIdentityResourceId) {
+        Objects.requireNonNull(managedIdentityResourceId);
+        this.inner.withAuthenticationType(AuthenticationType.MANAGED_IDENTITY.toString());
+        this.inner.withAdministratorLogin(managedIdentityResourceId);
+        // No administrator password is required for managed identity authentication.
+        this.inner.withAdministratorLoginPassword(null);
+        return this;
+    }
+
     SqlDatabaseExportRequestImpl withLoginAndPassword(String administratorLogin, String administratorPassword) {
         this.inner.withAdministratorLogin(administratorLogin);
         this.inner.withAdministratorLoginPassword(administratorPassword);