+++ sdk/sql/azure-resourcemanager-sql/src/test/java/com/azure/resourcemanager/sql/SqlServerTest.java
@@ -15,7 +15,6 @@ import com.azure.resourcemanager.resources.fluentcore.utils.ResourceManagerUtils
 import com.azure.resourcemanager.resources.ResourceManager;
 import com.azure.resourcemanager.storage.StorageManager;
 import com.azure.resourcemanager.test.ResourceManagerTestProxyTestBase;
-import com.azure.resourcemanager.test.model.AzureUser;
 import com.azure.resourcemanager.test.utils.TestDelayProvider;
 import com.azure.resourcemanager.test.utils.TestIdentifierProvider;
 
@@ -29,8 +28,6 @@ public abstract class SqlServerTest extends ResourceManagerTestProxyTestBase {
     protected StorageManager storageManager;
     protected String rgName = "";
     protected String sqlServerName = "";
-    private volatile String cachedAdminLogin;
-    private volatile String cachedAdminSid;
 
     @Override
     protected HttpPipeline buildHttpPipeline(TokenCredential credential, AzureProfile profile,
@@ -58,42 +55,4 @@ public abstract class SqlServerTest extends ResourceManagerTestProxyTestBase {
         resourceManager.resourceGroups().beginDeleteByName(rgName);
     }
 
-    /**
-     * Returns the signed-in user's principal name to use as the SQL Server Microsoft Entra administrator login.
-     * The Azure SQL platform policy requires Microsoft Entra-only authentication enabled at server creation time.
-     *
-     * @return the signed-in user's principal name (sanitized in playback)
-     */
-    protected String adminLogin() {
-        if (isPlaybackMode()) {
-            return "REDACTED";
-        }
-        ensureAdminIdentityCached();
-        return cachedAdminLogin != null ? cachedAdminLogin : "REDACTED";
-    }
-
-    /**
-     * Returns the signed-in user's object ID to use as the SQL Server Microsoft Entra administrator SID.
-     *
-     * @return the signed-in user's object ID (sanitized in playback)
-     */
-    protected String adminSid() {
-        if (isPlaybackMode()) {
-            return "00000000-0000-0000-0000-000000000000";
-        }
-        ensureAdminIdentityCached();
-        return cachedAdminSid != null ? cachedAdminSid : "00000000-0000-0000-0000-000000000000";
-    }
-
-    private synchronized void ensureAdminIdentityCached() {
-        if (cachedAdminLogin != null && cachedAdminSid != null) {
-            return;
-        }
-
-        AzureUser user = azureCliSignedInUser();
-        String login = user.userPrincipalName();
-        String id = user.id();
-        cachedAdminLogin = login != null ? login : "REDACTED";
-        cachedAdminSid = id != null ? id : "00000000-0000-0000-0000-000000000000";
-    }
 }